+++ src/main/java/com/bank_Project/bank_app/service/userService.java
@@ -4,9 +4,8 @@ import com.bank_Project.bank_app.DTO.UserDTO;
 
 public interface userService {
     public UserDTO insertUser(UserDTO userDTO);
-    public UserDTO deleteUser(int id);
-    public UserDTO getUserById(int id);
-    public  UserDTO updateUser(int id,UserDTO userDTO);
-
-    public String checkUser(int id);
+    public UserDTO deleteUser(Long id);
+    public UserDTO getUserById(Long id);
+    public  UserDTO updateUser(Long id,UserDTO userDTO);
+    public String checkUser(Long id);
 }
